requirements.txt
pyttsx3
numpy
scipy
sklearn
python_speech_features
SpeechRecognition
pyaudio
# Portaudio: On mac "brew install portaudio"
# For audio configuration, check here: https://pypi.org/project/SpeechRecognition/
pyyaml